Add tests for core methods and rename lall implementations.
The `lall` method in logpy.core could raise preventable EarlyExitError
if the caller didn't order the goals correctly. The old implementation of
lall has been renamed to lallgreedy (to reflect that it's a performance
optimization).

Note that the implementation of `earlyorder` is order preserving _if_
no goal raises an EarlyGoalError, so rename does not change the behavior
of `lall` in cases where goals do not need to be reordered.

There is some negative performance impact. However, the usability gain is
significant, and the old behavior is still available in lallgreedy if needed
for performance reasons.