Refactor `condp` and introduce `condpseq`
The `condp` operator is now more like the original version in the source paper.

Also, an experimental sequential-branch-state version has been added.  It uses
suggestion functions for each branch with a branch-cumulative state.  In other
words, it allows ordered branches to specify suggestion functions that can look
at the results of preceding states.